ci/requirements.txt
pyyaml
flake8
cmakelint
pycparser
pysoundfile
numpy
requests

